LayoutTests/editing/deleting/delete-emoji-2.html

<!DOCTYPE html>
<html>
<body>
<div id="test" contenteditable="true">&#x1f3f4;&#xe0067;&#xe0062;&#xe0065;&#xe006e;&#xe0067;&#xe007f;&#x1f3f4;&#xe0067;&#xe0062;&#xe0073;&#xe0063;&#xe0074;&#xe007f;&#x1f3f4;&#xe0067;&#xe0062;&#xe0077;&#xe006c;&#xe0073;&#xe007f;
</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
Markup.description("This test verifies that deletions are correct over emoji groups and emoji with variations");
var testElement = document.getElementById('test');
getSelection().setBaseAndExtent(testElement.firstChild, testElement.firstChild.length, testElement.firstChild, testElement.firstChild.length);
Markup.dump("test");
while (testElement.firstChild.length) {
    document.execCommand("Delete");
    Markup.dump("test");
}

</script>
</body>
</html>
