LayoutTests/platform/mac/editing/deleting/delete-emoji-2-expected.txt

This test verifies that deletions are correct over emoji groups and emoji with variations

Dump of markup 1:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤заЂДзаЂбзаЂ│заЂБзаЂ┤заЂ┐­ЪЈ┤заЂДзаЂбзаЂизаЂгзаЂ│заЂ┐<#selection-caret>
"

Dump of markup 2:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤заЂДзаЂбзаЂ│заЂБзаЂ┤заЂ┐­ЪЈ┤заЂДзаЂбзаЂизаЂгзаЂ│<#selection-caret>"

Dump of markup 3:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤заЂДзаЂбзаЂ│заЂБзаЂ┤заЂ┐­ЪЈ┤заЂДзаЂбзаЂизаЂг<#selection-caret>"

Dump of markup 4:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤заЂДзаЂбзаЂ│заЂБзаЂ┤заЂ┐­ЪЈ┤заЂДзаЂбзаЂи<#selection-caret>"

Dump of markup 5:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤заЂДзаЂбзаЂ│заЂБзаЂ┤заЂ┐­ЪЈ┤заЂДзаЂб<#selection-caret>"

Dump of markup 6:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤заЂДзаЂбзаЂ│заЂБзаЂ┤заЂ┐­ЪЈ┤заЂД<#selection-caret>"

Dump of markup 7:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤заЂДзаЂбзаЂ│заЂБзаЂ┤заЂ┐­ЪЈ┤<#selection-caret>"

Dump of markup 8:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤заЂДзаЂбзаЂ│заЂБзаЂ┤заЂ┐<#selection-caret>"

Dump of markup 9:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤заЂДзаЂбзаЂ│заЂБзаЂ┤<#selection-caret>"

Dump of markup 10:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤заЂДзаЂбзаЂ│заЂБ<#selection-caret>"

Dump of markup 11:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤заЂДзаЂбзаЂ│<#selection-caret>"

Dump of markup 12:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤заЂДзаЂб<#selection-caret>"

Dump of markup 13:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤заЂД<#selection-caret>"

Dump of markup 14:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐­ЪЈ┤<#selection-caret>"

Dump of markup 15:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂДзаЂ┐<#selection-caret>"

Dump of markup 16:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«заЂД<#selection-caret>"

Dump of markup 17:
| "­ЪЈ┤заЂДзаЂбзаЂЦзаЂ«<#selection-caret>"

Dump of markup 18:
| "­ЪЈ┤заЂДзаЂбзаЂЦ<#selection-caret>"

Dump of markup 19:
| "­ЪЈ┤заЂДзаЂб<#selection-caret>"

Dump of markup 20:
| "­ЪЈ┤заЂД<#selection-caret>"

Dump of markup 21:
| "­ЪЈ┤<#selection-caret>"

Dump of markup 22:
| <br>
